requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
decouple==0.0.7
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.16
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python-docx==0.8.11
python-docx-replace==0.4.4
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
